feat: save feed entry contents into topic contents
This change utilises the `content` property from rss-parser to populate the topic post content. That property contains HTML, so additional logic is present here to instruct the markdown plugin to ignore it. For versions of NodeBB that don't have an up-to-date Markdown plugin, the HTML will be parsed (and likely escaped), but admins can toggle "Allow HTML" to have it rendered properly.